requirements.txt
matplotlib
numpy
torch
numpy
Pillow
torch
torchvision
math
tqdm